feat(Products Tags): adding Tags to products and complete products implementation

+++ src/main/java/com/apogee/product/services/impl/ProductServiceImpl.java
@@ -1,6 +1,7 @@
 package com.apogee.product.services.impl;
 
 import com.apogee.product.entities.ProductEntity;
+import com.apogee.product.exceptions.RecordNotFoundException;
 import com.apogee.product.utilities.Mapper;
 import com.apogee.product.models.Product;
 import com.apogee.product.repositories.ProductRepository;
@@ -12,8 +13,8 @@ import org.springframework.stereotype.Service;
 import java.util.Collections;
 import java.util.List;
 import java.util.Optional;
-import java.util.concurrent.atomic.AtomicReference;
 
+import static com.apogee.product.utilities.Utilities.transform;
 import static com.apogee.product.utilities.Utilities.transformCollection;
 
 @Service
@@ -29,7 +30,8 @@ public class ProductServiceImpl implements ProductService {
         List<ProductEntity> productEntities = productRepository.findAll();
 
         if (!productEntities.isEmpty()) {
-            return transformCollection(productEntities, Product.class);
+
+            return transformCollection(productEntities, Product.class, this::setProductId);
         } else {
             return Collections.emptyList();
         }
@@ -39,32 +41,47 @@ public class ProductServiceImpl implements ProductService {
     public Product addProduct(Product product) throws Exception {
 
         ProductEntity transientProduct = Mapper.map(product, ProductEntity.class);
+
         ProductEntity savedEntity = productRepository.save(transientProduct);
 
-        return Mapper.map(savedEntity, Product.class);
+        return transform(savedEntity, Product.class, this::setProductId);
+    }
+
+    @Override
+    public Product updateProduct(Product product) throws Exception {
+
+        if (this.productRepository.existsById(product.getId())) {
+            
+            ProductEntity productEntity = Mapper.map(product, ProductEntity.class);
+
+            return transform(this.productRepository.save(productEntity), Product.class, this::setProductId);
+        } else {
+            throw new RecordNotFoundException("record.not.found", product.getId());
+        }
     }
 
     @Override
     public Product findProductById(Long productId) throws Exception {
 
-        AtomicReference<Product> productReference = new AtomicReference<>();
         Optional<ProductEntity> productEntityOptional = this.productRepository.findById(productId);
 
-        productEntityOptional.ifPresent(productEntity -> {
-            try {
-                productReference.set(Mapper.map(productEntity, Product.class));
-            } catch (Exception e) {
-                throw new RuntimeException(e);
-            }
-        });
-
-        return productReference.get();
+        return transform(productEntityOptional.orElseThrow(() -> new RecordNotFoundException("record.not.found", productId)), Product.class, this::setProductId);
     }
 
     @Override
     public void deleteProductById(Long productId) throws Exception {
 
-        this.productRepository.deleteById(productId);
+        if (this.productRepository.existsById(productId)) {
+
+            this.productRepository.deleteById(productId);
+        } else {
+            throw new RecordNotFoundException("record.not.found", productId);
+        }
+    }
+
+    private Product setProductId(ProductEntity categoryEntity, Product product) {
+        product.setId(categoryEntity.getId());
+        return product;
     }
